[CORE] turned item size into a double

+++ core/core/src/main/java/hu/zagor/gamebooks/character/handler/item/DefaultCharacterItemHandler.java
@@ -67,8 +67,8 @@ public class DefaultCharacterItemHandler implements CharacterItemHandler {
 
     private boolean canAddItem(final Character character, final Item item) {
         final int backpackSize = character.getBackpackSize();
-        int itemsInBackpack = 0;
-        int itemSize = 0;
+        double itemsInBackpack = 0;
+        double itemSize = 0;
 
         if (itemIsInBackpack(item)) {
             for (final Item ownedItem : character.getEquipment()) {
